Add startup notification timestamp hack
In the case that we are invoked with a DISPLAY but not a
DESKTOP_STARTUP_ID, "steal" the current timestamp out of the server and
act like it's the one we were invokved with.

This deals with the case of launching gedit from a terminal and
expecting the existing window to raise itself.