Restore proper handling of workspaces
Activate an existing window only if it is on the current workspace. "Current" here is
defined as "current for the X server"... there is a theoretical race because we ask
what is the "current" workspace when we receive the message on the primary instance,
but that's not a problem and the X reply is async anyway.