search and replace: use async functions
Async search has a price: more lines of code...

The forward and backward search can be done via the replace dialog, or
via the menu. In the first case, when the operation is finished, a
message must be displayed in the statusbar and the sensitivity of the
replace button must be updated. That's why there are two different
callbacks: from or not from the dialog window.