Fix sensitivity of search menu actions
For the find prev, find next and clear highlight menu actions. They
relied on the "search-text" buffer property. But now the "search-text"
property is in the SearchSettings. And the SearchSettings is accessible
via a SearchContext. And GeditDocument can have two different
SearchContext (from GeditViewFrame and from the search and replace).

So it is a little more difficult to implement.