Show regex error in the search and replace dialog
The error appears as a symbolic icon in the GtkEntry (secondary icon, on
the right). The error message is in the tooltip.

The secondary icon is better in my opinion, because it doesn't move the
text while we are typing it.

Depending on the regex error type, the error icon is shown in the search
entry, or the replace entry.