Fix escape bug with the search and replace completion
The completion contents was unescaped before inserting it, and escaped
after retrieving it. But the unescape/escape functions are not
reciprocal. So it's easier to always deal with the original text.

As a consequence, the gedit_history_entry_set_escape_func() function is
no longer needed.

Little problem: there will be bugs with old completion entries, when
upgrading the gedit version.