Search: use the new API for the regex replace errors
When a regex replace error occurs, the error icon with the tooltip is
added to the replace entry. The error icon is removed when the
replacement text changes. The new replacement text can still contain
errors, but it will be reported only when the user clicks again on the
replace button.

And when there is a replace error, the replace button is insensitive.