Replace dialog: create the SearchContext before the actual search
The object method handler gedit_replace_dialog_response() creates the
SearchContext if it is not already created.

In commands-search.c, replace_dialog_response_cb() must be called after
the SearchContext creation, to execute the actual find or replace.

If replace_dialog_response_cb() is called before, the SearchContext may
not be already created, and the find or replace doesn't work.