Replace dialog: do not bind the search settings
The reason of the previous behavior is now explained in the code.

The regex errors appear when a button is pressed, and disappear when the
search entry's text is modified. Before, a regex compilation error was
reported directly, as soon as the pattern is typed. This is possible if
we create a hidden SearchContext (without highlighting) besides the
normal SearchContext. This hidden SearchContext would be created
directly, and also on each document tab switch.