Load highlight mode from metadata for *.txt files
guess_language() is called when the content type is modified. For a
*.txt file, the content type has the same value as the init value, so
the notify callback is not called, and the highlight mode is not
changed. To fix the problem, we also call guess_language() when the
document is loaded.

For other files than *.txt, guess_language() is thus called two times in
loaded(): when the content type is modified (via the notify callback),
and when guess_language() is called directly.

But it's better two times than zero.