CloseConfirmationDialog: UI improvements
The dialog is now resizable on multi mode. It is useful when the list of
unsaved documents is long, and we want to show all of them.

The scrolled window around the tree view has a greater minimum height,
so by default more documents can be shown.

The vbox2 must have the expand and fill properties to true (with
gtk_box_pack_start()), so the treeview expands when the dialog window is
resized.

The size request of the tree view is removed, since it has no effect
inside a scrolled window (but it's maybe a bug in gtk, I just get used
to work around the problem, and set the size on the scrolled window
instead).

(and the policy for a scrolled window is "automatic" by default in GTK+ 3)

It would have been better to have (1) natural height if below N (2)
scrolled window if above N. With N at a treeview row boundary. But it is
not easily feasible with gtk+ afaik (in gtksourceview we use a custom
container for that...).

https://bugzilla.gnome.org/show_bug.cgi?id=708711