Allow plugins to extend the appmenu
Rename the API to just be extend_menu instead of extend_gear_menu
and look for the requested extension point in all the menus we have.
Plugins are updated for the API change and snippets and tools now
put their pref dialogs in the app menu.
We still need to rename all the extension points to something more
meaningful of "ext1" and we probably should add similar API to the
app-actiavatable.