Move menu extension to the app-activatable
It is important that the menu-model for the gear menu is a single
instance shared among all the windows. This makes it consistent
with the app-menu model and plugins can extend them in the same
way without causing items to be added multiple times when there
is more than a window.
The extension API is thus moved to the app activatable interface
and the plugins are adapted to use the new api.