gedit-app: fix gedit --wait
Fix a longstanding bug in "wait mode" so that the invoked instance
always exits when the tab is closed, even in the case that it was the
first-run of gedit and another document was opened since.

We do this by forcing D-Bus activation (using IS_LAUNCHER flag), so -w
might not work if gedit is not installed.