replace-dialog: fix corner-case button sensitivity
If the regular expression checkbutton is checked, the search entry
contains an invalid regular expression, and the user presses Find or
Replace All, all response buttons correctly become insensitive. Now, if
the user unchecks the regex checkbutton, the error should be cleared and
the responses should become sensitive so that he can search again. This
wasn't happening.

https://bugzilla.gnome.org/show_bug.cgi?id=726593