W32: load libgedit.dll at runtime
Since libgedit is linked as delay-loaded DLL, it's still not loaded when
main() is called. Before we call functions from it, we must locate
libgedit.dll and load it (otherwise, if it's not in PATH (which is how it
should be!), calling its functions will crash the program).
As comments indicate, the use of g_module_open() is not really necessary,
but, given the circumstances, it is convenient and clean.

https://bugzilla.gnome.org/show_bug.cgi?id=727663