Do not remove missing remote files from the recent list
Failure to load a remote file is often transient so we prefer
to keep such file in the recent list even if load/save fails.
While at it split recent file utils to a separate file: they were
in window because in the past access to the recent manager required
a reference to the current screen.