Do not run the document loader/saver unit tests with make check
It is difficult to configure the environment to run those unit tests:
gvfs, ftp, even some values in GSettings, etc.

It is anyway planned in the near future to move the document
loader/saver in GtkSourceView, where the environment will be simpler to
configure to run the unit tests (no file metadata with gvfs, and no
GSettings configuration required).

The unit tests are simply not run for "make check". But they can still
be run separately.