File browser: focus the text cell before editing
https://bugzilla.gnome.org/show_bug.cgi?id=697101

This patch sets the focus cell to the text cell, as it was not
possible to start a rename before clicking an item in the tree
view because of the focus cell being the icon cell, which can
not be edited.
After setting the focus cell to the text cell, new file and folder
creation starts a rename even before the first tree selection.