Don't leak the last GeditView on application shutdown
At first I thought all GeditView were leaked, so I searched where a ref
of GeditView is not released, and I've finally found.

But after some testing, only the last GeditView of the application is
leaked, because an unref in GIO must be executed in an idle, and the
idle is not executed on application shutdown.

So even if the leak was not really a leak, this commit will make
debugging memory leaks a bit nicer, and will hopefully save some time
for next developers.