Simplify window actions sensitivity handling
Centralize everything in single function to be called every time
something that affects sensitivity changes.
This is not only much shorter and simpler, but also more correct
since we had actions ending up with the wrong sensitivity after
different update functions were called.
I do not think this has performance problems (as a matter of fact
I think we may be doing less toggling in practice) but if it ever
turns out to be too slow we could pass in a hint of what are the
actions affected by a change.