EncodingsComboBox: update correctly the menu
The menu was updated *before* the settings are saved, so the menu was
incorrectly updated. Now the operations are executed in the correct
order: first save the gsetting, then update the combo box menu.

Now the dialog window is destroyed externally (by the combo box),
instead of internally (which is quite strange). So that the
connect_after() works.