Let Ctrl+W close the window if there are no tabs
gedit is one of very few apps where Ctrl+W, Ctrl+Q and Alt+F4 are all necessary
to use the app.  In particular, Alt+F4 is the only way to get rid of a window
if i press Ctrl+N instead of Ctrl+T.

An extra Ctrl+W press could close the window after closing the last tab.  It is
easy for the user to understand what happened, and once you know it you don't
have to learn another shortcut.

It turns out that this feature was actually already supported, but only on OS
X.  In _gedit_cmd_file_close the #ifdef is useless, since it will never be
reached when tab != NULL disables the action.  The patch then only has to
remove that condition from update_actions_sensitivity.

It makes sense now to remove the #ifdef from update_actions_sensitivity
too, but which branch to take?  As requested by jessevdk, I kept the
non-OS X branch; the remaining conditions where Ctrl+W is disabled are:

- closing/saving/print-previewing, where I'm not really sure how to test
it and they should be rare anyway.  Perhaps for saving it should set a
close-after-save flag?

- for print preview, I'd find it confusing that you'll close the entire
document instead of just the preview

- for saving error, it's better to get the user's attention

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>

https://bugzilla.gnome.org/show_bug.cgi?id=735514