commands-file: async/finish version of gedit_commands_save_document()
The important part is the finish(), to get the result. Because
gedit_commands_save_document() was already async, but it was difficult
to know whether the document was correctly saved (by listening to the
GeditDocument signals it _was_ possible, but it's no longer possible
because the error parameter has been removed from the "saved" signal).

https://bugzilla.gnome.org/show_bug.cgi?id=739686