Port the external-tools plugin to the new file saving API
Listening to the "saving" and "saved" signals, with a counter to know if
all documents are correctly saved, was quite hackish. The new gedit API
is cleaner, with GIO conventions for async operations.

Anyway the port was needed because the code was broken. The "saving"
signal has been removed, and the error parameter of the "saved" signal
has been removed.

https://bugzilla.gnome.org/show_bug.cgi?id=739686