Bump API version
gedit 3.14 broke the API, and we forgot to bump the version.

Also, move the variable to the top, so we'll hopefully not forget it in
the future (unless we endlessly put things to the top when we forget
something and the variable is again at the bottom when we need it).