Window: do not allow saving when the print preview is shown
Supporting document saving when the print preview is shown is not really
useful and makes the code in GeditTab more complicated and error-prone.
A later commit will simplify the code in GeditTab.

What will still be supported is to quit gedit and save unsaved
documents, even if for some of them the print preview is shown. In that
case, the print preview will simply be closed before saving.

Also, the Save As was not well supported with the print preview, the
file name was not updated in the print preview.