GeditDocument: add property use-gvfs-metadata
The goal is to make the gedit-document code reusable for other apps.
Later, it'll be easier if we don't depend too much on config.h.

From config.h, only ENABLE_GVFS_METADATA is used in gedit-document.
Create a property for that setting, and configure it at only one place,
in gedit_document_new().

In the future, a new subclass of GtkSourceBuffer can be created, e.g.
GcodeDocument, and GeditDocument can be a subclass of GcodeDocument.
GcodeDocument should not depend on config.h, so that the code can be
copied more easily in other apps without the need to add build time
configuration.