GeditDocument: bind syntax-highlighting gsetting
The goal is to concentrate the use of GSettings in one place in
GeditDocument, so that it'll be easier to make the code reusable.

Having the "highlight-syntax" property as TRUE with a NULL language is
fine for GtkSourceBuffer, so the code is easier like that. A binding is
also better, so if the gsetting changes, it is reflected directly in
gedit.