Deprecate unused utils functions and macros (wip)
Maybe third-party plugins still use those deprecated functions. In that
case the function can be copied into the plugin, or the plugin author
can contact the gedit developers to undeprecate the function (if there
is a good reason to do so).

For macros a deprecated guard is needed for GTK-Doc, so add the
GEDIT_DISABLE_DEPRECATED for the deprecated guard.