notebook: use gtk_notebook_get_tab_label()
It is not needed to set the tab-label GObject data, because we can use
gtk_notebook_get_tab_label().

But we need to disconnect the signal in gedit_notebook_remove(), before
the page removal (that is, before the chain-up), so we can still
retrieve the tab_label. In gedit_notebook_page_removed(), the page is
already removed, so it's not possible to retrieve the tab_label there.