tab: remove GTask instance variable
Instead of storing the GTask into a GeditTab instance variable, pass the
GTask as a parameter to the relevant functions. GTask is intended to be
used like that.

This has the advantage that the file saving is now more independent and
could be extracted into its own class in the future, if GeditTab
provides the needed APIs (possibly private APIs).