tab: use a GTask for the file loading and reverting
To have less object attributes in the GeditTab struct.

The change is only internal to GeditTab. The code in gedit doesn't need
the result of the file loading or reverting, so the async/finish
functions are static.

It is possible in the future that the file loading code will be
delegated to another class. Using a GTask will ease that goal.