Move tab DND support to GeditNotebook
To remove the dependency on GeditNotebook in GeditView. A class should
not be aware of who contains it. The purpose is to make GeditView more
self-contained, so it can be reused more easily in other text editors.

Instead of handling the DND in the object method handler of GeditView,
we connect to the signal externally, which works fine too. And since the
code is related to notebooks, it's more logical to have that code in
GeditNotebook.