Update view centering in GeditWindow
The view centering was updated in GeditViewFrame (depending on whether
the GtkWindow is in fullscreen mode). A class should not be aware of who
contains it.

This commit moves the code to GeditWindow. GeditViewFrame is thus more
self-contained and more reusable. The fullscreen mode is an attribute of
the window, so it's logical to put the code in GeditWindow (to keep
related data and behavior together).

The code in GeditWindow is also more straightforward, no need to
disconnect/reconnect signals etc.