print-preview: simplify set_zoom_fit_to_size()
Call set_zoom_factor() instead of computing ourself the tile size.
set_zoom_factor() calls update_tile_size(), which is equivalent to the
code that updated manually the tile size, except when

width - 2 * PAGE_PAD
or
height - 2 * PAGE_PAD

is negative (in the case of a very small gedit window). In that case,
the new code is slightly better since it takes 1 pixel to draw the page.
But the new code is better in the sense that update_tile_size() can be
modified later to get_tile_size() instead of storing the values.