print-preview: get tile size with a getter
No need to store the value, since it can be computed from other values.

It is not a good idea to store duplicated data. Performances aside, only
the minimum amount of data should be stored. With duplicated data, we
must pay attention to not have inconsistencies.

Here the performances should not matter, the getter is a O(1) function.