spell-plugin: improve memory management for the auto-spell
Create a ViewData struct and attach it to the view. The struct is freed
when the view is destroyed, which unrefs the auto_spell.

In ViewData, the GeditSpellPlugin instance is needed, to access the
"auto-spell" GAction of the window. If it was not needed, then
implementing GeditViewActivatable would have been easier.