auto-spell: fix critical message about setting tag priority
The critical message was:
Gtk-CRITICAL **: gtk_text_tag_set_priority: assertion 'priv->table != NULL' failed

In tag_removed_cb() we need to check if the tag removed is the
tag_highlight.

Also, the doc of gtk_text_tag_set_priority() says that the tag has the
highest priority by default, just after being added to the TagTable, so
no need to set the priority in that case.