spell: factor out a SpellNavigator
The idea is to make more code reusable. gedit-spell-plugin is specific
to gedit.

GeditSpellNavigator is an interface.
GeditSpellNavigatorGtv implements the interface for a GtkTextView.

A later commit will remove the SpellCheckerDialog signals, and use
directly a SpellNavigator. That way, to support other types of buffers
(e.g. for WebKitGTK+), one just need to implement a new SpellNavigator.
Implementing a new spell checker widget is also easy, for example
instead of a dialog window, one can implement a new widget that could be
packed below an headerbar.

The UI and the logic are separated.