spell-checker-dialog: show message "No misspelled words"
It was before shown in the statusbar, but showing it in the dialog is
also a good solution. Ideally the dialog should have an "empty state",
or "completed state", by hiding the widgets and show only the message
with an icon. In the statusbar it is easier to miss the message.