gedit-recent: casefold/normalize utf8 strings before comparing them
The recent file list is currently checked against the open document
selector search entry by converting both the uri and the user input
to lowercase.

This isn't ideal, because the same lowercase string can be represented
using different valid byte sequences, and because there isn't a
one-to-one mapping of uppercase and lowercase letters.

This commit changes the logic to instead normalize the string into
the standard reduced form, and use utf8 case folding to be more robust
and various inputs.

https://bugzilla.gnome.org/show_bug.cgi?id=753481