Move minimap settings in the preferences dialog
We decided to keep this in the preferences dialog for now instead
of adding yet another item to the menu.
In the UI we refer to the minimap with the term "Overview Map", which
is what Builder uses.
This also changes the setting name and position to be more consistent
with other settings (line numbers, right margin, etc)