GeditDocument: make sure that metadata_info exists
To fix this warning:
CRITICAL **: set_gvfs_metadata: assertion 'G_IS_FILE_INFO (info)' failed

The bug occurred when opening a non-existing file from the command line,
and changing something (e.g. enable the spell checking) that sets a
metadata.

An empty GFileInfo is used to collect the metadata even if the file
doesn't exist. So when the file is saved, the metadata isn't lost.