Refer to the BugTracker gedit wiki page
Instead of directly linking to bugzilla, add an indirection to the wiki
page that contains a list of frequent bugs and some instructions, so
hopefully we'll receive less duplicated bug reports, with better
descriptions.

Remove the BUGS file, since the information is already present in the
README, on the main wiki page, in the man page, in the doap, etc.

Do not change the URL in the doap, since the doap is parsed by some
tools, and they maybe assume it's a link to bugzilla.