app: remove some deprecated gdk_screen_* methods
This code impact the tab vs new window opening policy,
some cleanup is done but it's still not working under Wayland
as Gdk is returning (0,0) for the window position.

Under X11, the same policy on x coordinates
(in the viewport if 25% of the window is in it)
is used for the Y coordinates.