Revert "Add shortcut for inserting a timestamp"
This reverts commit 90e92d79e5b2a031a675c2aebc3a3b457fd6a445.

I feel really bad about doing this, but I think we should
be a bit careful about adding a new keyboard shortcut right
at the Freeze. I don't object to there being a shortcut for this;
I just want us to pick the right shortcut.

Let's land this or something similar early in the next release
cycle and do an earlier release to have more chance for feedback.