readme: big update, simplify
Throughout the file: better use the Markdown syntax (GitLab renders it).

In order in the diff:

Intro:
- Take the same description as in the AppData file. There is anyway more
information on the website. To not have many different gedit
descriptions a bit everywhere (more difficult to update at all the
places).
- For the programming languages used, it's explained in
  docs/gedit-development-getting-started.md.

Resources:
- all the links are available on the website.

Installation:
- Have a separate Dependencies section, and refer to the meson.build
file, it will simplify the maintenance, to have the list of deps in only
one place, no need to update the version numbers in both places (in the
past the readme file was frequently outdated, so it was worse).
- Simpler install procedure with Meson, with the 3 commands taken from:
  https://mesonbuild.com/Quick-guide.html
- Don't explain how to extract the tarball, if someone wants to compile
gedit from source, the person hopefully knows that it's needed to
extract the tarball first (also to be able to read the readme file).

Contributions:
- Don't duplicate infos from the hacking file.