Remove CONTRIBUTING.md, outdated and no longer needed
It explained to send patches on bugzilla! We are now on GitLab.

And there is now a script dealing with pull requests on the GitHub
mirror of GNOME repos, explaining that it is not the right place where
to contribute.

Note that there is of course the HACKING file that explains how to
contribute.