Remove MAINTAINERS file, no longer needed, see the *.doap file
And it was outdated, Jeremy Bicha was not present in MAINTAINERS but is
present in gedit.doap.

Normally the MAINTAINERS file is no longer needed for Gnome projects.
Every service that needs the list of maintainers read the *.doap file
now (e.g. l10n.gnome.org). Also, there are other projects like gspell
that don't contain a MAINTAINERS file since the beginning and nobody
complained :-) So I guess it's fine.