contributing: refer to the GtkSourceView HACKING file
Add a link to the GtkSourceView HACKING file, and don't duplicate its
content in the gedit CONTRIBUTING file. It's a big simplification. In
the past I've synced the two files between GSV and gedit, then I've
improved the GSV HACKING file but I was no longer contributing to gedit
so the gedit file was not updated; now it's fixed.

The "Building from git" section is already explained in the
`docs/gedit-development-getting-started.md` file.