Remove GeditViewCentering
Centering the text view was done to have a more distraction-free UX. But
it was not well integrated into gedit: it was enabled only on fullscreen
AND if the right margin was set, without the possibility to disable it
in that case. For users it's difficult to know what's going on, whether
the right margin is set the text editor behaves differently for an
unrelated action. Users have asked to have more options to
enable/disable the feature.

To re-enable the feature to center the view, it should be implemented in
a better way:
1) Wrap GeditView into a GeditViewCentering container only when needing
to center the text. In other words, if the text is not centered, then
GeditViewCentering should not be used at all. So, the
GeditView/GtkScrolledWindow should be dynamically added/removed from the
GeditViewCentering.
2) In the preferences dialog, provide options to enable/disable the
feature for the normal/fullscreen modes.

It's also a pragmatic decision to simplify the gedit code, because there
are not enough developers, and too many bugs. Do you prefer a rock-solid
text editor? Or a text editor with some shiny and non-essential features
that complexifies the code, at the expense of more bugs? The gedit core
should have a much smaller code size, with all the essential features
implemented in libraries.

-----

diff details: it's GeditViewCentering that created the
GtkScrolledWindow, so in the *.ui file we need to create the
GtkScrolledWindow instead of GeditViewCentering.