ViewFrame: disable overlay scrolling for the GtkScrolledWindow
The overlay scrollbars get in the way when trying to place the text
cursor with the mouse. The problem occurs with both the right-side
scrollbar and the bottom scrollbar. It's especially annoying when text
wrapping is disabled, to place the cursor at the last line. An ugly
workaround was to place the cursor at a different character, then move
the cursor with the keyboard. It's a GTK regression that was never
properly fixed in gedit.