Document: use TeplFileMetadata to re-implement set/get_metadata()
TeplFile and TeplFileMetadata are used only internally by GeditDocument.
The implementation is similar to the previous implementation, but it's
shorter, and GeditMetadataManager has been removed since it's present in
Tepl.

Tepl is smarter and automatically falls back to the metadata manager in
case GVfs metadata gives an error. But it's configurable with a build
option.