Remove overview map
It currently doesn't work well with the additional normal scrollbar that
is always displayed (see commit 5e0909c19f267b4023131881fac89c58a2a17453
for why it is needed).

It's also a pragmatic decision to simplify the gedit code, because there
are not enough developers, and too many bugs. Do you prefer a rock-solid
text editor? Or a text editor with some shiny and non-essential features
that complexifies the code, at the expense of more bugs? The gedit core
should have a much smaller code size, with all the essential features
implemented in libraries.