FileChooserDialogGtk: revert to previous implementation
Revert "file chooser: use g_object_unref to free a native dialog"
This reverts commit 967b27be72d6ede8db272f885b91124bd6d45194.

Revert "gedit: convert "All Text Files" filter to use a fixed set of mime types"
This reverts commit 7c5c36efc7943d3df51a904cf11fcb02ab03e751.

Revert "gedit: don't assume file chooser is a GtkWindow"
This reverts commit 9ce64b92ffd7c3c5d1a858c0eef20ab3203b97f5.

Revert "gedit: port GeditFileChooserDialogGtk to GtkFileChooserNative"
This reverts commit 34419a44bc9f7186278e540d335b7d1b17f54dae.

GeditFileChooserDialogGtk was ported to GtkFileChooserNative in the hope
that gedit could be more sandboxed with Flatpak and Snap. But access to
the host file system is still needed for the integrated file browser in
the side panel.

Also, more importantly, when GeditFileChooserDialogGtk was ported to
GtkFileChooserNative, an essential feature was removed: the ability to
choose any character encoding. In the combobox to choose the character
encoding when opening a file, there is now again the "Add or Remove..."
to configure the list.

There are plans in Tepl to avoid the need to have a combobox in the open
file chooser dialog, and provide the choice for the character encoding
elsewhere in the application. But this requires more work.